Create README.md
This LWC allows you to plug the icon design library directly into a screen flow. So you can call any icon of any type and variant into your screenflow. This is a small QoL improvement to help make the end experience a little more polished. This will be coming to an org near you soon when this project releases.